+++ spec/integration/sinatra/sinatra_spec.rb
@@ -1,6 +1,10 @@
 require 'sinatra'
 require 'spec_helper'
+
 require 'apps/sinatra/dummy_app'
+require 'apps/sinatra/composite_app/sinatra_app1'
+require 'apps/sinatra/composite_app/sinatra_app2'
+
 require 'integration/shared_examples/rack_examples'
 
 RSpec.describe "Sinatra integration specs" do
@@ -12,6 +16,62 @@ RSpec.describe "Sinatra integration specs" do
     it "includes version" do
       get '/crash'
       wait_for_a_request_with_body(/"context":{.*"version":"1.2.3 Sinatra/)
+    end
+  end
+
+  context "when multiple apps are mounted" do
+    let(:endpoint1) do
+      'https://airbrake.io/api/v3/projects/113743/notices?key=fd04e13d806a90f96614ad8e529b2822'
+    end
+
+    let(:endpoint2) do
+      'https://airbrake.io/api/v3/projects/99123/notices?key=ad04e13d806a90f96614ad8e529b2821'
+    end
+
+    def env_for(url, opts = {})
+      Rack::MockRequest.env_for(url, opts)
+    end
+
+    before do
+      stub_request(:post, endpoint1).to_return(status: 201, body: '{}')
+      stub_request(:post, endpoint2).to_return(status: 201, body: '{}')
+    end
+
+    context "and when both apps use their own notifiers and middlewares" do
+      let(:app) do
+        Rack::Builder.new do
+          map('/app1') do
+            use Airbrake::Rack::Middleware, SinatraApp1
+            run SinatraApp1.new
+          end
+
+          map '/app2' do
+            use Airbrake::Rack::Middleware, SinatraApp2
+            run SinatraApp2.new
+          end
+        end
+      end
+
+      it "reports errors from SinatraApp1 notifier" do
+        get '/app1'
+
+        body = %r|"backtrace":\[{"file":".+apps/sinatra/composite_app/sinatra_app1.rb"|
+
+        wait_for(
+          a_request(:post, endpoint1).
+            with(body: body)
+        ).to have_been_made.once
+      end
+
+      it "reports errors from SinatraApp2 notifier" do
+        get '/app2'
+
+        body = %r|"backtrace":\[{"file":".+apps/sinatra/composite_app/sinatra_app2.rb"|
+        wait_for(
+          a_request(:post, endpoint2).
+            with(body: body)
+        ).to have_been_made.once
+      end
     end
   end
 end